requirements.txt
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pykafka==2.8.0
